\name{nlnc-package}
\alias{nlnc-package}
\alias{nlnc}
\docType{package}
\title{
\packageTitle{nlnc}
}
\description{
\packageDescription{nlnc}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{nlnc}
\packageIndices{nlnc}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{nlnc}

Maintainer: \packageMaintainer{nlnc}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
